refactor: split CLI into pluggable sub-command architecture
- Extract display.py (Rich styles, color_level, console)
- Extract utils.py (load_entries, compute_overview)
- Move log commands to commands/logs.py (logs_app)
- main.py becomes thin orchestrator — one add_typer() per group
- Update poetry.lock with typer/rich dependencies

Adding new command groups requires 3 steps:
1. Create commands/<group>.py with a Typer sub-app
2. Add app.add_typer(<group>_app) to main.py
3. Share display/utils helpers across commands
