+++ backend/src/ala/services/project_manager.py
@@ -215,7 +215,9 @@ _ALA_MARKERS: list[str] = [
 def _is_ala_root(path: Path) -> bool:
     """Return True when *path* looks like an ALA project root."""
     # Must have pyproject.toml
-    if not (path / "pyproject.toml" if not (path / "backend" / "pyproject.toml").exists() else True):
+    if not (
+        path / "pyproject.toml" if not (path / "backend" / "pyproject.toml").exists() else True
+    ):
         return False
     # At least one ALA marker must exist
     for marker in _ALA_MARKERS:
